Fix makie extensions
